requirements.txt
psycopg2-binary
pandas
boto3
jupyter
ipython-sql


